backend/requirements.txt
# ===== Core API =====
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-socketio
aiohttp

# ===== Machine Learning - Attention Detection =====
torch
torchvision
timm
numpy
pillow

# ===== Audio Transcription =====
openai-whisper
soundfile
librosa

# ===== AI Assistant com LangChain =====
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38

# Document Loaders - Universal (substitui pypdf, python-docx, openpyxl, python-pptx)
unstructured[all-docs]

# Vector Store - FAISS
faiss-cpu==1.8.0

# Embeddings - Sentence Transformers
sentence-transformers==2.7.0

# Text Processing
tiktoken==0.6.0
